Move comment related components under comments container
